All notice codes removed; cleanup and tests still pending
`DefaultObjectManager::set_entry` now signals a failed operation via a thrown exception instead of a notice code.

All notice code values have been removed. The enum itself still exists, and traces of it and `detailed_notice` are lurking around in various places, yet to be removed.

The files for form data no longer include `notice_code_list.h`.

Removed member functions (on the `file_save` interface and on `file_load_order`) for logging file save warnings as `detailed_notice` instances.

Removed the unused frontend type `multithreadable_load_results`.

+++ native/src/DovahKit/dovah/notices/file_load_errors/base_subrecord_load_error.h
@@ -9,7 +9,7 @@ namespace dovah::notices::file_load_errors {
 
          struct {
             uint32_t signature = 0;
-            uint16_t size      = 0;
+            uint32_t size      = 0;
 
             uint32_t offset_into_record = 0;
          } subrecord;
